Split EdgeCachedMapper out of graph partitioner
